moved all styles into their own package
