+++ src/main/java/Styles/BoardStyle.java
@@ -1,4 +1,4 @@
-package main;
+package Styles;
 
 import java.awt.*;
