+++ src/main/java/Styles/DefaultBoardStyle.java
@@ -1,22 +1,23 @@
-package main;
+package Styles;
 
 import java.awt.*;
 
 /**
- * A dark mode style for the mancala board
+ * The default style for the mancala board which uses brown colors to match
+ * that of a wooden board.
  * @author Jonathan Stewart Thomas
  * @version 1.0.0.230426
  */
-public class DarkModeStyle implements BoardStyle{
+public class DefaultBoardStyle implements BoardStyle{
     private final Color boardColor, fontColor, pitColor;
 
     /**
-     * Sets the colors for the dark style
+     * Sets the colors for the default style
      */
-    public DarkModeStyle() {
-        boardColor = new Color(0, 0, 0);
-        fontColor = new Color(255,255,255);
-        pitColor = new Color(73, 73, 73);
+    public DefaultBoardStyle() {
+        boardColor = new Color(163, 114, 46);
+        fontColor = new Color(0,0,0);
+        pitColor = new Color(133, 91, 17);
     }
 
     /**
